Move etcd watch code to its own files
