+++ pkg/tools/etcd_tools.go
@@ -20,12 +20,8 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"sync"
 
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/watch"
 	"github.com/coreos/go-etcd/etcd"
-	"github.com/golang/glog"
 )
 
 const (
@@ -299,322 +295,5 @@ func (h *EtcdHelper) AtomicUpdate(key string, ptrToType interface{}, tryUpdate E
 			continue
 		}
 		return err
-	}
-}
-
-// FilterFunc is a predicate which takes an API object and returns true
-// iff the object should remain in the set.
-type FilterFunc func(obj interface{}) bool
-
-// Everything is a FilterFunc which accepts all objects.
-func Everything(interface{}) bool {
-	return true
-}
-
-// WatchList begins watching the specified key's items. Items are decoded into
-// API objects, and any items passing 'filter' are sent down the returned
-// watch.Interface. resourceVersion may be used to specify what version to begin
-// watching (e.g., for reconnecting without missing any updates).
-func (h *EtcdHelper) WatchList(key string, resourceVersion uint64, filter FilterFunc) (watch.Interface, error) {
-	w := newEtcdWatcher(true, filter, h.Codec, h.ResourceVersioner, nil)
-	go w.etcdWatch(h.Client, key, resourceVersion)
-	return w, nil
-}
-
-// Watch begins watching the specified key. Events are decoded into
-// API objects and sent down the returned watch.Interface.
-func (h *EtcdHelper) Watch(key string, resourceVersion uint64) (watch.Interface, error) {
-	return h.WatchAndTransform(key, resourceVersion, nil)
-}
-
-// WatchAndTransform begins watching the specified key. Events are decoded into
-// API objects and sent down the returned watch.Interface. If the transform
-// function is provided, the value decoded from etcd will be passed to the function
-// prior to being returned.
-//
-// The transform function can be used to populate data not available to etcd, or to
-// change or wrap the serialized etcd object.
-//
-//   startTime := time.Now()
-//   helper.WatchAndTransform(key, version, func(input interface{}) (interface{}, error) {
-//     value := input.(TimeAwareValue)
-//     value.Since = startTime
-//     return value, nil
-//   })
-//
-func (h *EtcdHelper) WatchAndTransform(key string, resourceVersion uint64, transform TransformFunc) (watch.Interface, error) {
-	w := newEtcdWatcher(false, Everything, h.Codec, h.ResourceVersioner, transform)
-	go w.etcdWatch(h.Client, key, resourceVersion)
-	return w, nil
-}
-
-// TransformFunc attempts to convert an object to another object for use with a watcher
-type TransformFunc func(interface{}) (interface{}, error)
-
-// etcdWatcher converts a native etcd watch to a watch.Interface.
-type etcdWatcher struct {
-	encoding  Codec
-	versioner ResourceVersioner
-	transform TransformFunc
-
-	list   bool // If we're doing a recursive watch, should be true.
-	filter FilterFunc
-
-	etcdIncoming  chan *etcd.Response
-	etcdStop      chan bool
-	etcdCallEnded chan struct{}
-
-	outgoing chan watch.Event
-	userStop chan struct{}
-	stopped  bool
-	stopLock sync.Mutex
-
-	// Injectable for testing. Send the event down the outgoing channel.
-	emit func(watch.Event)
-}
-
-// newEtcdWatcher returns a new etcdWatcher; if list is true, watch sub-nodes.  If you provide a transform
-// and a versioner, the versioner must be able to handle the objects that transform creates.
-func newEtcdWatcher(list bool, filter FilterFunc, encoding Codec, versioner ResourceVersioner, transform TransformFunc) *etcdWatcher {
-	w := &etcdWatcher{
-		encoding:      encoding,
-		versioner:     versioner,
-		transform:     transform,
-		list:          list,
-		filter:        filter,
-		etcdIncoming:  make(chan *etcd.Response),
-		etcdStop:      make(chan bool),
-		etcdCallEnded: make(chan struct{}),
-		outgoing:      make(chan watch.Event),
-		userStop:      make(chan struct{}),
-	}
-	w.emit = func(e watch.Event) { w.outgoing <- e }
-	go w.translate()
-	return w
-}
-
-// etcdWatch calls etcd's Watch function, and handles any errors. Meant to be called
-// as a goroutine.
-func (w *etcdWatcher) etcdWatch(client EtcdGetSet, key string, resourceVersion uint64) {
-	defer util.HandleCrash()
-	defer close(w.etcdCallEnded)
-	if resourceVersion == 0 {
-		latest, ok := etcdGetInitialWatchState(client, key, w.list, w.etcdIncoming)
-		if !ok {
-			return
-		}
-		resourceVersion = latest
-	}
-	_, err := client.Watch(key, resourceVersion, w.list, w.etcdIncoming, w.etcdStop)
-	if err != etcd.ErrWatchStoppedByUser {
-		glog.Errorf("etcd.Watch stopped unexpectedly: %v (%#v)", err, key)
-	}
-}
-
-// etcdGetInitialWatchState turns an etcd Get request into a watch equivalent
-func etcdGetInitialWatchState(client EtcdGetSet, key string, recursive bool, incoming chan<- *etcd.Response) (resourceVersion uint64, success bool) {
-	success = true
-
-	resp, err := client.Get(key, false, recursive)
-	if err != nil {
-		if !IsEtcdNotFound(err) {
-			glog.Errorf("watch was unable to retrieve the current index for the provided key: %v (%#v)", err, key)
-			success = false
-			return
-		}
-		if index, ok := etcdErrorIndex(err); ok {
-			resourceVersion = index + 1
-		}
-		return
-	}
-	resourceVersion = resp.EtcdIndex + 1
-	convertRecursiveResponse(resp.Node, resp, incoming)
-	return
-}
-
-// convertRecursiveResponse turns a recursive get response from etcd into individual response objects
-// by copying the original response.  This emulates the behavior of a recursive watch.
-func convertRecursiveResponse(node *etcd.Node, response *etcd.Response, incoming chan<- *etcd.Response) {
-	if node.Dir {
-		for i := range node.Nodes {
-			convertRecursiveResponse(node.Nodes[i], response, incoming)
-		}
-		return
-	}
-	copied := *response
-	copied.Action = "get"
-	copied.Node = node
-	incoming <- &copied
-}
-
-// translate pulls stuff from etcd, convert, and push out the outgoing channel. Meant to be
-// called as a goroutine.
-func (w *etcdWatcher) translate() {
-	defer close(w.outgoing)
-	defer util.HandleCrash()
-
-	for {
-		select {
-		case <-w.etcdCallEnded:
-			return
-		case <-w.userStop:
-			w.etcdStop <- true
-			return
-		case res, ok := <-w.etcdIncoming:
-			if !ok {
-				return
-			}
-			w.sendResult(res)
-		}
-	}
-}
-
-func (w *etcdWatcher) decodeObject(data []byte, index uint64) (interface{}, error) {
-	obj, err := w.encoding.Decode(data)
-	if err != nil {
-		return nil, err
-	}
-
-	// ensure resource version is set on the object we load from etcd
-	if w.versioner != nil {
-		if err := w.versioner.SetResourceVersion(obj, index); err != nil {
-			glog.Errorf("failure to version api object (%d) %#v: %v", index, obj, err)
-		}
-	}
-
-	// perform any necessary transformation
-	if w.transform != nil {
-		obj, err = w.transform(obj)
-		if err != nil {
-			glog.Errorf("failure to transform api object %#v: %v", obj, err)
-			return nil, err
-		}
-	}
-
-	return obj, nil
-}
-
-func (w *etcdWatcher) sendCreate(res *etcd.Response) {
-	if res.Node == nil {
-		glog.Errorf("unexpected nil node: %#v", res)
-		return
-	}
-	data := []byte(res.Node.Value)
-	obj, err := w.decodeObject(data, res.Node.ModifiedIndex)
-	if err != nil {
-		glog.Errorf("failure to decode api object: '%v' from %#v %#v", string(data), res, res.Node)
-		// TODO: expose an error through watch.Interface?
-		// Ignore this value. If we stop the watch on a bad value, a client that uses
-		// the resourceVersion to resume will never be able to get past a bad value.
-		return
-	}
-	if !w.filter(obj) {
-		return
-	}
-	action := watch.Added
-	if res.Node.ModifiedIndex != res.Node.CreatedIndex {
-		action = watch.Modified
-	}
-	w.emit(watch.Event{
-		Type:   action,
-		Object: obj,
-	})
-}
-
-func (w *etcdWatcher) sendModify(res *etcd.Response) {
-	if res.Node == nil {
-		glog.Errorf("unexpected nil node: %#v", res)
-		return
-	}
-	curData := []byte(res.Node.Value)
-	curObj, err := w.decodeObject(curData, res.Node.ModifiedIndex)
-	if err != nil {
-		glog.Errorf("failure to decode api object: '%v' from %#v %#v", string(curData), res, res.Node)
-		// TODO: expose an error through watch.Interface?
-		// Ignore this value. If we stop the watch on a bad value, a client that uses
-		// the resourceVersion to resume will never be able to get past a bad value.
-		return
-	}
-	curObjPasses := w.filter(curObj)
-	oldObjPasses := false
-	var oldObj interface{}
-	if res.PrevNode != nil && res.PrevNode.Value != "" {
-		// Ignore problems reading the old object.
-		if oldObj, err = w.decodeObject([]byte(res.PrevNode.Value), res.PrevNode.ModifiedIndex); err == nil {
-			oldObjPasses = w.filter(oldObj)
-		}
-	}
-	// Some changes to an object may cause it to start or stop matching a filter.
-	// We need to report those as adds/deletes. So we have to check both the previous
-	// and current value of the object.
-	switch {
-	case curObjPasses && oldObjPasses:
-		w.emit(watch.Event{
-			Type:   watch.Modified,
-			Object: curObj,
-		})
-	case curObjPasses && !oldObjPasses:
-		w.emit(watch.Event{
-			Type:   watch.Added,
-			Object: curObj,
-		})
-	case !curObjPasses && oldObjPasses:
-		w.emit(watch.Event{
-			Type:   watch.Deleted,
-			Object: oldObj,
-		})
-	}
-	// Do nothing if neither new nor old object passed the filter.
-}
-
-func (w *etcdWatcher) sendDelete(res *etcd.Response) {
-	if res.PrevNode == nil {
-		glog.Errorf("unexpected nil prev node: %#v", res)
-		return
-	}
-	data := []byte(res.PrevNode.Value)
-	obj, err := w.decodeObject(data, res.PrevNode.ModifiedIndex)
-	if err != nil {
-		glog.Errorf("failure to decode api object: '%v' from %#v %#v", string(data), res, res.PrevNode)
-		// TODO: expose an error through watch.Interface?
-		// Ignore this value. If we stop the watch on a bad value, a client that uses
-		// the resourceVersion to resume will never be able to get past a bad value.
-		return
-	}
-	if !w.filter(obj) {
-		return
-	}
-	w.emit(watch.Event{
-		Type:   watch.Deleted,
-		Object: obj,
-	})
-}
-
-func (w *etcdWatcher) sendResult(res *etcd.Response) {
-	switch res.Action {
-	case "create", "get":
-		w.sendCreate(res)
-	case "set", "compareAndSwap":
-		w.sendModify(res)
-	case "delete":
-		w.sendDelete(res)
-	default:
-		glog.Errorf("unknown action: %v", res.Action)
-	}
-}
-
-// ResultChannel implements watch.Interface.
-func (w *etcdWatcher) ResultChan() <-chan watch.Event {
-	return w.outgoing
-}
-
-// Stop implements watch.Interface.
-func (w *etcdWatcher) Stop() {
-	w.stopLock.Lock()
-	defer w.stopLock.Unlock()
-	// Prevent double channel closes.
-	if !w.stopped {
-		w.stopped = true
-		close(w.userStop)
 	}
 }
